requirements.txt
# UI and App Framework
streamlit>=1.32.0

# OpenAI (or Azure/OpenAI or DeepSeek — choose based on your use)
openai>=1.30.1
# OR use this if you're using Azure OpenAI:
# azure-ai-openai>=1.0.0b1

# LangChain for chaining LLMs and using DuckDuckGo search tool
langchain>=0.1.20
langchain-community>=0.0.35

# DuckDuckGo search tool
duckduckgo-search>=4.0

# Wikipedia integration
wikipedia>=1.4.0

# Environment variable loading
python-dotenv>=1.0.1

# Password hashing for login system
bcrypt>=4.1.2

# Optional — For API calls to DeepSeek or other models
requests>=2.31.0

streamlit-cookies-manager>=0.1.0

